+++ example/lib/main.adapter.g.m8.dart
@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// Emitted on: 2019-05-16 02:23:34.320239
+// Emitted on: 2019-05-16 12:47:12.529398
 
 // **************************************************************************
 // DatabaseProviderGenerator